+++ core/types/aliases.h
@@ -14,13 +14,20 @@ namespace Fresa
 
     //---String---
     using str = std::string;
-    inline std::vector<str> split(str s, str del = " ") {
+    inline void lower(str &s) {
+        std::transform(s.begin(), s.end(), s.begin(), [](unsigned char c){ return std::tolower(c); });
+    }
+    inline std::vector<str> split(str s, str del = " ", bool recursive = false) {
         std::vector<str> ss{};
         int a = 0;
         int b = (int)s.find(del);
         while (b != -1) {
             ss.push_back(s.substr(a, b - a));
             a = (int)(b + del.size());
+            if (recursive) {
+                while (s.size() > a and s.substr(a, del.size()) == del)
+                    a++;
+            }
             b = (int)s.find(del, a);
         }
         ss.push_back(s.substr(a, b - a));